Chal1/requirements.txt
# Python Dependencies for AeroEyes Challenge

# Core libraries
numpy>=1.21.0
scipy>=1.7.0
opencv-python>=4.5.0
Pillow>=9.0.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# YOLOv8
ultralytics>=8.0.0

# Data augmentation
albumentations>=1.3.0

# Data processing
pandas>=1.3.0
scikit-learn>=1.0.0
PyYAML>=6.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0

# Model optimization
onnx>=1.12.0
onnxruntime>=1.13.0
# onnxruntime-gpu>=1.13.0  # Optional: For GPU inference (comment out if not needed)

# TensorRT (Jetson only - install separately on device)
# tensorrt>=8.5.0
# pycuda>=2021.1

# Export formats (optional - requires cmake)
# onnxsim>=0.4.0  # Requires cmake: sudo apt-get install cmake

# Utilities
python-dateutil>=2.8.0
pytz>=2021.3

# Optional: Web API
# flask>=2.0.0
# requests>=2.26.0
